refactor(user): 컨트롤러 역할별 분리
785줄 단일 파일을 3개로 분리
- controller/user.controller.ts — 회원가입, Google OAuth, 내 정보
- controller/user-admin.controller.ts — 승인·거절·목록·편집
- controller/user-class-rep.controller.ts — 반대표 승인·목록
