+++ src/user/user.module.ts
@@ -3,15 +3,17 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserService } from './service/user.service';
 import { UserAdminService } from './service/user-admin.service';
 import { UserClassRepService } from './service/user-class-rep.service';
-import { UserController } from './user.controller';
+import { PermissionService } from './service/permission.service';
+import { UserController } from './controller/user.controller';
+import { UserAdminController } from './controller/user-admin.controller';
+import { UserClassRepController } from './controller/user-class-rep.controller';
 import { User } from './user.entity';
 import { StudentClassModule } from '../student-class/student-class.module';
-import { PermissionService } from './service/permission.service';
 import { GoogleModule } from '../google/google.module';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User]), StudentClassModule, GoogleModule],
-  controllers: [UserController],
+  controllers: [UserController, UserAdminController, UserClassRepController],
   providers: [UserService, UserAdminService, UserClassRepService, PermissionService],
   exports: [UserService, UserAdminService, UserClassRepService, PermissionService],
 })
